server: split join_list to its own file
Previously StoreSpec and JoinListType were in the same file. This
commit moves JoinListType and its test to a separate file.

Epic: CRDB-41111

Release note: None
