+++ pkg/base/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "cluster_id.go",
         "config.go",
         "constants.go",
+        "join_list.go",
         "license.go",
         "node_id.go",
         "store_spec.go",
@@ -53,6 +54,7 @@ go_test(
         "addr_validation_test.go",
         "cluster_id_test.go",
         "config_test.go",
+        "join_list_test.go",
         "main_test.go",
         "node_id_test.go",
         "store_spec_test.go",
